Enable passing objects as values for individual namespaces
In some extensions there is a need exclude some namespaces for some
options (ex. option is not applicable to Talk namespaces). Also, some
options come with preset namespace, for which value must not be changed.

This implementation enabled passing object instead of just bool as the
value for individiual NS.
[
   'value' => true/false,
   'disable' => true/false,
   'read_only' => true/false
]

Read-only values will be displayed, but cannot be changed, while
disabled values will display '-' icon in the panel and will be hidden
from the edit dialog

Change-Id: I2d2d5793a2a8f41f26c69e225c1eb796c8cbfe9d