Force content namespace for NS_MAIN + fix subpages
NS_MAIN should not be removed as content namespace

Explicitly set false if subpages are disabled for a NS

Remove namespace columns from tests, as they might leave the backend as
arrays. In the grid they will get converted to boolean, so type is
correct for that. If we want to make it unified, we should change all
extensions attaching to NSM to use arrays

Change-Id: I2b2d931e6f394dd55d0502bdf450119d750e30c7
ERM: #10762