Fix creating alias with illegal characters
After some namespace was edited all other namespaces which were edited at least once (have some settings in "nm-settings.php" file) got their aliases.
And if there is no predefined alias - alias is set from title.
Which is wrong, for example in case with NS_MAIN namespace.
It's title is "(Pages)", and it can't be alias because of parenthesises.

[3.2.x]

ERM24379

Change-Id: Ia5f1020d112c8bfd7f8acc3f8e65acc889ce86e5